requirements.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
python-docx==0.8.11
PyPDF2==3.0.1
requests==2.31.0
